users, tasks, subscribers

+++ src/api/store/team.py
@@ -1,6 +1,7 @@
 from _main_.utils.footage.FootageConstants import FootageConstants
 from _main_.utils.footage.spy import Spy
 from api.tests.common import RESET
+from api.utils.api_utils import is_admin_of_community
 from api.utils.filter_functions import get_team_member_filter_params, get_teams_filter_params
 from database.models import Team, UserProfile, Media, Community, TeamMember, CommunityAdminGroup, UserActionRel
 from _main_.utils.massenergize_errors import MassEnergizeAPIError, InvalidResourceError, CustomMassenergizeError, NotAuthorizedError
@@ -215,8 +216,7 @@ class TeamStore:
           (ADMIN_URL_ROOT, team.id))
 
         for cadmin in cadmins:
-          send_massenergize_email(subject="New team awaiting approval",
-                                msg=message, to=cadmin.email)
+          send_massenergize_email(subject="New team awaiting approval",msg=message, to=cadmin.email)
       team.save()
       for admin in verified_admins:
         teamMember, _ = TeamMember.objects.get_or_create(team=team,user=admin)
@@ -346,9 +346,18 @@ class TeamStore:
     try:
       team_id = args["id"]
       teams = Team.objects.filter(id=team_id)
+
       if not teams:
         return None, InvalidResourceError()
-
+      
+      #  is context user an admin of the primary community?
+      if context.user_is_community_admin and not is_admin_of_community(context, teams.first().primary_community.id):
+        return None, NotAuthorizedError()
+      
+      team_member = TeamMember.objects.filter(team=teams.first(), user=context.user_id).first()
+      
+      if not team_member.is_admin:
+        return None, NotAuthorizedError()
 
       # team.members deprecated.  Delete TeamMembers separate step
       team = teams.first()
@@ -366,13 +375,12 @@ class TeamStore:
       return None, CustomMassenergizeError(e)
 
 
-  def join_team(self, args) -> Tuple[Team, MassEnergizeAPIError]:
+  def join_team(self,context, args) -> Tuple[Team, MassEnergizeAPIError]:
     try:
       team_id = args.get("id", None)
-      user_id = args.get("user_id", None)
 
       team = Team.objects.get(id=team_id)
-      user = UserProfile.objects.get(id=user_id)
+      user = UserProfile.objects.get(id=context.user_id)
       teamMember, created = TeamMember.objects.get_or_create(team=team, user=user)
       if created:
         teamMember.save()
@@ -382,12 +390,12 @@ class TeamStore:
       capture_message(str(e), level="error")
       return None, CustomMassenergizeError(e)
 
-  def leave_team(self, args) -> Tuple[Team, MassEnergizeAPIError]:
+  def leave_team(self, context,args) -> Tuple[Team, MassEnergizeAPIError]:
     try:
       team_id = args.get("id", None)
       user_id = args.get("user_id", None)
       team = Team.objects.get(id=team_id)
-      user = UserProfile.objects.get(id=user_id)
+      user = UserProfile.objects.get(id=context.user_id)
       teamMember = TeamMember.objects.filter(team=team, user=user)
       if teamMember:
         teamMember.delete()
@@ -407,6 +415,9 @@ class TeamStore:
       is_admin = args.get("is_admin", False)
 
       team = Team.objects.get(id=team_id)
+
+      if context.user_is_community_admin and not is_admin_of_community(context, team.primary_community.id):
+          return None, CustomMassenergizeError('You are not authorized to add members to this team')
 
       if user_id:
         user = UserProfile.objects.get(id=user_id)
@@ -437,6 +448,9 @@ class TeamStore:
       elif email:
         user = UserProfile.objects.get(email=email)
 
+      if context.user_is_community_admin and not is_admin_of_community(context, team.primary_community.id):
+          return None, CustomMassenergizeError('You are not authorized to remove members from this team')
+
       team_member = TeamMember.objects.filter(team__id=team_id, user=user)
       if team_member.count() > 0:
         team_member.delete()
@@ -459,6 +473,11 @@ class TeamStore:
       team_id = args.get('team_id', None)
       if not team_id:
         return [], CustomMassenergizeError('Please provide a valid team_id')
+      
+      team = Team.objects.filter(id=team_id).first()
+      # for cadmins, allow only if admin of teams parent community
+      if context.user_is_community_admin and not is_admin_of_community(context, team.primary_community.id):
+          return None, CustomMassenergizeError('You are not authorized to view members of this team')
 
       members = TeamMember.objects.filter(is_deleted=False, team__id=team_id, user__accepts_terms_and_conditions=True, user__is_deleted=False, *filter_params)
       return members.distinct(), None
@@ -479,9 +498,14 @@ class TeamStore:
         # only list users that have joined the platform
         if user.accepts_terms_and_conditions:
           member = TeamMember.objects.filter(user=user, team=team).first()
-          member_obj = {"id": None, "user_id": str(user.id), "preferred_name": user.preferred_name, "is_admin": False}
+          member_obj = {
+            # "id": None,
+            # "user_id": str(user.id),
+            "preferred_name": user.preferred_name, 
+            "is_admin": False
+            }
           if member:
-            member_obj['id'] = member.id
+            # member_obj['id'] = member.id 
             member_obj['is_admin'] = member.is_admin
           res.append(member_obj)
 
@@ -509,17 +533,16 @@ class TeamStore:
         return teams, None
 
       community_id = args.pop('community_id', None)
-      if community_id == 0:
-        # return actions from all communities
-        return self.list_teams_for_super_admin(context)
-      
-      elif not community_id:
+    
+      if not community_id:
         user = UserProfile.objects.get(pk=context.user_id)
         admin_groups = user.communityadmingroup_set.all()
         comm_ids = [ag.community.id for ag in admin_groups]
         teams = Team.objects.filter(communities__id__in = comm_ids, is_deleted=False, *filter_params).select_related('logo', 'primary_community')
         return teams.distinct(), None
-
+      
+      if context.user_is_community_admin and not is_admin_of_community(context, community_id):
+          return None, CustomMassenergizeError('You are not authorized to view members of this team')
       teams = Team.objects.filter(communities__id=community_id, is_deleted=False,*filter_params).select_related('logo', 'primary_community')   
       return teams.distinct(), None
